+++ CHANGELOG.md
@@ -14,6 +14,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - version 1.0.46
 
 ### Added
+- add getAllScheduledActivities for OFSC activities and bump version to 1.0.56
 - add getAllNonScheduledActivities function
 - Bump version to 1.0.52 and add activity lifecycle actions
 - update version to 1.0.51 and add user/inventory type management functions